--- lib/Utilities/constants.dart
@@ -1,4 +1,0 @@
-import 'package:flutter/material.dart';
-
-Color primaryColor = Color(0xFF333333);
-Color secondaryColor = Color(0xFFf37121);
