+++ lib/core/constants.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+Color primaryColor = Color(0xFF333333);
+Color secondaryColor = Color(0xFFf37121);
