+++ lib/presentation/Widgets/view/custom_bottomsheet.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:openscan/Utilities/constants.dart';
-import 'package:openscan/Utilities/database_helper.dart';
+import 'package:openscan/core/constants.dart';
+import 'package:openscan/core/data/database_helper.dart';
 import 'package:openscan/presentation/Widgets/delete_dialog.dart';
 import 'package:openscan/presentation/screens/home_screen.dart';
 import 'package:openscan/presentation/screens/view_document.dart';
