+++ lib/presentation/Widgets/view/quality_selector.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:openscan/Utilities/constants.dart';
+import 'package:openscan/core/constants.dart';
 
 class QualitySelector extends StatefulWidget {
   final int imageQuality;
