+++ lib/presentation/screens/about_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:openscan/Utilities/constants.dart';
+import 'package:openscan/core/constants.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 void launchWebsite(String urlString) async {
